+++ src/contexts/dataContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+import { IDataContext } from '../utils/interfaces';
+
+export const DataContext = createContext<IDataContext>({
+  data: null,
+  setData: () => null,
+});
