+++ src/contexts/searchContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+import { ISearchContext } from '../utils/interfaces';
+
+export const SearchContext = createContext<ISearchContext>({
+  searchValue: '',
+  setSearchValue: () => null,
+});
